pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terag"
version = "0.6.0"
description = "Token-Efficient Retrieval Augmented Generation with Graph-based Indexing"
readme = "README.md"
authors = [
    { name = "Rudra Naik", email = "rudra91@gmail.com" }
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "networkx",
    "scikit-learn",
    "openai",
    "tiktoken",
    "groq",
    "python-dotenv",
    "tqdm"
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort"
]
fuzzy = [
    "fuzzywuzzy",
    "python-Levenshtein"
]

[project.urls]
"Homepage" = "https://github.com/rudranaik/terag"
"Bug Tracker" = "https://github.com/rudranaik/terag/issues"

[tool.setuptools.packages.find]
where = ["terag"]
